+++ core/src/main/java/org/apache/calcite/rel/logical/LogicalWindow.java
@@ -103,6 +103,11 @@ public final class LogicalWindow extends Window {
       getRowType(), groups);
   }
 
+  @Override public Window copy(List<RexLiteral> constants) {
+    return new LogicalWindow(getCluster(), getTraitSet(), getInput(),
+        constants, getRowType(), groups);
+  }
+
   /**
    * Creates a LogicalWindow.
    *